# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
        "/example/helloworld": {
            "get": {
                "description": "do ping",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "example"
                ],
                "summary": "ping example",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/registration/customers": {
            "get": {
                "description": "Responds with the list of all customers as JSON.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "Get customers array",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.customer"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Takes a customer JSON and store in DB. Return saved JSON.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "newcustomer"
                ],
                "summary": "Store a new customer",
                "parameters": [
                    {
                        "description": "customer Data",
                        "name": "customer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.customer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.customer"
                        }
                    }
                }
            }
        },
        "/registration/customers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "getCustomerByID locates the customer whose ID value matches the id",
                "operationId": "get-customer-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "customer ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.customer"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.customer": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "countrycode": {
                    "type": "integer"
                },
                "customertype": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phonenumber": {
                    "type": "integer"
                },
                "verified": {
                    "type": "string"
                }
            }
        },
        "main.message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}